Feature lwdev 4903 fees in common cash operations

+++ contract/Lykke.Service.OperationsRepository.Contract/Abstractions/IBaseCashOperation.cs
@@ -22,7 +22,7 @@ namespace Lykke.Service.OperationsRepository.Contract.Abstractions
 
         bool IsHidden { get; }
 
-        decimal FeeSize { get; }
+        double FeeSize { get; }
 
         FeeType FeeType { get; }
     }
